Varias Cosas

+++ UI-Base-De-Datos/src/app/Shared/Card-Reply/Card-Reply.component.ts
@@ -1,3 +1,4 @@
+import { mostrarError } from 'Domain/GlobalErrorHandler';
 import { Component, Input, OnInit } from '@angular/core';
 import { Replica } from 'Domain/Replica';
 import { ReplyService } from 'src/app/Services/Reply.service';
@@ -19,6 +20,7 @@ export class CardReplyComponent implements OnInit {
   editing:boolean = false
   faTrashCan=faTrashCan
   faPen=faPen
+  errors = []
 
   constructor(private replyService:ReplyService) { }
 
@@ -45,11 +47,23 @@ export class CardReplyComponent implements OnInit {
   }
 
   async deleteReply(){
-    console.log(this.replica.id)
-    await this.replyService.deleteReply(this.replica.id)
-    const index = this.replyList.indexOf(this.replica, 0);
-      if (index > -1) {
-    this.replyList.splice(index, 1);
+    if(this.replica.replicas.length == 0){
+      await this.replyService.deleteReply(this.replica.id)
+      const index = this.replyList.indexOf(this.replica, 0);
+        if (index > -1) {
+      this.replyList.splice(index, 1);
+      }
+    }else{
+      mostrarError(this,new Error("No se puede borrar una replica con replicas"))
     }
   }
+
+  buttonText(){
+    if(this.isOn){
+      return "Ocultar Replicas"
+    }else{
+      return "Mostrar Replicas"
+    }
+  }
+
 }
